[misc] Rename paginate component
